feat(spectral): Spectral validation in the browser
This moves the spectral validation from a backend service to the client
js bundle, hopefully speeding up the validation process.

The change will allow us to drop the second container running the
spectral API as we will now only need to ship the static files to the
browser, which can be done by the already bundled nginx.

Files are expected to be placed under the /spectral directory in a structure
like this:

```
├── apic-v10.yml
├── apic.yml
└── spectral
   ├── apic-v10.yml
   ├── apic.yml
   ├── apisyouwonthate.yml
   └── functions
      ├── rule-a.js
      └── rule-b.js
```

The downside of this approach is a significantly larger client bundle,
which is why the size checks in webpack were turned from an error to a warning.

Implements HSP-5882