feat(spectral): Cancel in-flight fetch calls
We might have in-flight validation attempts
that need to be cancelled when a new one is about to be issued
or we risk getting the same issue displayed multiple times
if the timing is exactly right/wrong and both complete after all `arg.errActions.clear` have run.

To work around this we use an AbortController, fetch has built-in support for that.
Sadly the easy solution of globally creating a controller and then simply sending an abort()
on every run of validateSpec will immediately cancel the new fetch that is supposed to happen.

To work around this a new AbortController is constructed on every run of validateSpec and stored in the controller variable.
The controller variable is null at start and will be set back to null after every fetch() call.
If it isn't null when validateSpec is called then this must mean that a request is currently going on and that can be cancelled.
Overriding the controller afterwards with a new AbortController will avoid the signal also cancelling the new fetch() call